Move mreg unit tests where do they belong
